notifications permission before entering app for first time
